refactor(commerce): simple business order card
