+++ apps/qopnet-commerce/components/index.ts
@@ -6,6 +6,7 @@ export * from './header'
 export * from './home'
 export * from './products'
 export * from './profile'
+export * from './simple-business-order'
 export * from './supplier-product'
 export * from './supplier'
 export * from './upload'
